+++ Modules/Intent.Modules.Blazor/Templates/Templates/Client/RazorComponentStyle/RazorComponentStyleTemplate.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -13,13 +14,19 @@ namespace Intent.Modules.Blazor.Templates.Templates.Client.RazorComponentStyle
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public override string TransformText()
         {
-            return StyleContentOverride ?? "";
+            if (!string.IsNullOrWhiteSpace(StyleContentOverride))
+            {
+                return StyleContentOverride;
+            }
+
+            var filePath = GetMetadata().GetFilePath();
+            return System.IO.File.Exists(filePath) ? System.IO.File.ReadAllText(filePath) : "";
         }
 
         [IntentManaged(Mode.Ignore)]
         public override bool CanRunTemplate()
         {
-            return !string.IsNullOrWhiteSpace(StyleContentOverride);
+            return !string.IsNullOrWhiteSpace(StyleContentOverride) || System.IO.File.Exists(GetMetadata().GetFilePath());
         }
     }
 }